+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-n/a
+### Added
+- Introduced ``Client.get_content_type_ids`` method to retrieve supported content types.
 
 ## [1.3.0] - 2019-08-15